Rename sort hook file
