+++ packages/components/src/List/ListComposition/Manager/ListManager.component.js
@@ -5,7 +5,7 @@ import { withTranslation } from 'react-i18next';
 import { ListContext } from '../context';
 import getDefaultT from '../../../translate';
 import I18N_DOMAIN_COMPONENTS from '../../../constants';
-import useCollectionSort from './hooks/sort.hook';
+import useCollectionSort from './hooks/useCollectionSort.hook';
 import { filterCollectionByText } from './filter';
 import useLocalStorage from './hooks/useLocalStorage.hook';
